+++ CHANGELOG.md
@@ -8,6 +8,8 @@
 * [`PowPersistentSession.Plug.Base`] Now includes `:pow_config` in the store config
 * [`PowResetPassword.Plug`] Now includes `:pow_config` in the store config
 * [`Pow.Plug.Base`] Now includes `:pow_config` in the store config
+* [`Pow.Operations`] Added `Pow.Operations.reload/2` to reload structs
+* [`PowPersistentSession.Store.PersistentSessionCache`] Update `PowPersistentSession.Store.PersistentSessionCache.get/2` to reload the user using `Pow.Operations.reload/2`
 
 ## v1.0.21 (2020-09-13)